chore: consolidate EnumElim macro into IntroElim.
