+++ utilities/proc-macros/src/intro_elim.rs
@@ -14,11 +14,21 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use proc_macro2::TokenStream;
-use quote::quote;
+use proc_macro2::{Ident, TokenStream};
+use quote::{format_ident, quote};
 use syn::{Data, DeriveInput, Fields};
 
 pub fn derive(input: DeriveInput) -> TokenStream {
+	match &input.data {
+		Data::Struct(_) => derive_struct(input),
+		Data::Enum(_) => derive_enum(input),
+		Data::Union(_) =>
+			syn::Error::new_spanned(&input.ident, "IntroElim cannot be derived for unions")
+				.to_compile_error(),
+	}
+}
+
+fn derive_struct(input: DeriveInput) -> TokenStream {
 	let name = &input.ident;
 	let generics = &input.generics;
 	let fields = match &input.data {
@@ -32,10 +42,7 @@ pub fn derive(input: DeriveInput) -> TokenStream {
 				.to_compile_error();
 			},
 		},
-		Data::Enum(_) | Data::Union(_) => {
-			return syn::Error::new_spanned(name, "IntroElim can only be derived for structs")
-				.to_compile_error();
-		},
+		Data::Enum(_) | Data::Union(_) => unreachable!(),
 	};
 
 	let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
@@ -59,4 +66,77 @@ pub fn derive(input: DeriveInput) -> TokenStream {
 			}
 		}
 	}
+}
+
+fn derive_enum(input: DeriveInput) -> TokenStream {
+	let enum_ident = input.ident;
+	let generics = input.generics;
+	let variants = match input.data {
+		Data::Enum(data) => data.variants,
+		Data::Struct(_) | Data::Union(_) => unreachable!(),
+	};
+
+	let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
+	let output = format_ident!("__IntroElimOutput");
+
+	let handler_args = variants.iter().enumerate().map(|(variant_index, variant)| {
+		let handler = handler_ident(variant_index);
+		let field_types = variant.fields.iter().map(|field| &field.ty);
+		quote! {
+			#handler: impl Fn(#(#field_types),*) -> #output
+		}
+	});
+
+	let match_arms = variants.iter().enumerate().map(|(variant_index, variant)| {
+		let variant_ident = &variant.ident;
+		let handler = handler_ident(variant_index);
+		let field_bindings = (0..variant.fields.len())
+			.map(|field_index| field_ident(variant_index, field_index))
+			.collect::<Vec<_>>();
+
+		match &variant.fields {
+			Fields::Named(fields) => {
+				let patterns =
+					fields.named.iter().zip(field_bindings.iter()).map(|(field, binding)| {
+						let field_ident =
+							field.ident.as_ref().expect("named fields have identifiers");
+						quote! { #field_ident: #binding }
+					});
+				quote! {
+					Self::#variant_ident { #(#patterns),* } => #handler(#(#field_bindings),*)
+				}
+			},
+			Fields::Unnamed(_) => {
+				quote! {
+					Self::#variant_ident(#(#field_bindings),*) => #handler(#(#field_bindings),*)
+				}
+			},
+			Fields::Unit => {
+				quote! {
+					Self::#variant_ident => #handler()
+				}
+			},
+		}
+	});
+
+	quote! {
+		impl #impl_generics #enum_ident #ty_generics #where_clause {
+			pub fn elim<#output>(
+				self,
+				#(#handler_args,)*
+			) -> #output {
+				match self {
+					#(#match_arms,)*
+				}
+			}
+		}
+	}
+}
+
+fn handler_ident(index: usize) -> Ident {
+	format_ident!("__elim_handler_{}", index)
+}
+
+fn field_ident(variant_index: usize, field_index: usize) -> Ident {
+	format_ident!("__elim_{}_{}", variant_index, field_index)
 }
