+++ utilities/proc-macros/src/lib.rs
@@ -18,7 +18,6 @@ use proc_macro::TokenStream;
 
 mod arbitrary;
 mod better_modules;
-mod enum_elim;
 mod generate_module;
 mod intro_elim;
 mod type_introspection;
@@ -137,12 +136,6 @@ pub fn generate_module(_attr: TokenStream, item: TokenStream) -> TokenStream {
 pub fn derive_has_type_introspection(input: TokenStream) -> TokenStream {
 	let input = syn::parse_macro_input!(input as syn::DeriveInput);
 	type_introspection::derive(input).into()
-}
-
-#[proc_macro_derive(EnumElim)]
-pub fn derive_enum_elim(input: TokenStream) -> TokenStream {
-	let input = syn::parse_macro_input!(input as syn::DeriveInput);
-	enum_elim::derive(input).into()
 }
 
 #[proc_macro_derive(IntroElim)]
